chore: simplify view for events IT and move createEventString helper to super class
