+++ vaadin-charts-flow-parent/vaadin-charts-flow-demo/src/main/java/com/vaadin/flow/component/charts/demo/AbstractChartExample.java
@@ -5,8 +5,17 @@ import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.ScheduledFuture;
 import java.util.concurrent.TimeUnit;
 
+import com.fasterxml.jackson.annotation.JsonAutoDetect;
+import com.fasterxml.jackson.annotation.JsonIgnoreType;
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.PropertyAccessor;
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.SerializationFeature;
 import com.vaadin.flow.component.Component;
+import com.vaadin.flow.component.ComponentEvent;
 import com.vaadin.flow.component.UI;
+import com.vaadin.flow.component.charts.Chart;
 import com.vaadin.flow.component.html.Div;
 import com.vaadin.flow.server.Command;
 
@@ -22,6 +31,29 @@ public abstract class AbstractChartExample extends Div {
 
     protected void showNotification(String message) {
         UI.getCurrent().getPage().executeJs("window.alert($0);", message);
+    }
+
+    protected String createEventString(ComponentEvent<Chart> event) {
+        ObjectMapper mapper = new ObjectMapper()
+                .setSerializationInclusion(JsonInclude.Include.NON_NULL)
+                .disable(SerializationFeature.FAIL_ON_EMPTY_BEANS)
+                .enable(SerializationFeature.INDENT_OUTPUT)
+                .setVisibility(PropertyAccessor.ALL,
+                        JsonAutoDetect.Visibility.NONE)
+                .setVisibility(PropertyAccessor.FIELD,
+                        JsonAutoDetect.Visibility.ANY)
+                .addMixIn(Command.class, JacksonMixinForIgnoreCommand.class);
+
+        try {
+            return mapper.writeValueAsString(event);
+        } catch (JsonProcessingException e) {
+            e.printStackTrace();
+            return "";
+        }
+    }
+
+    @JsonIgnoreType
+    static class JacksonMixinForIgnoreCommand {
     }
 
     /**
