+++ vaadin-charts-flow-parent/vaadin-charts-flow-demo/src/main/java/com/vaadin/flow/component/charts/demo/examples/dynamic/ServerSideEvents.java
@@ -1,16 +1,7 @@
 package com.vaadin.flow.component.charts.demo.examples.dynamic;
 
-import com.fasterxml.jackson.annotation.JsonAutoDetect;
-import com.fasterxml.jackson.annotation.JsonIgnoreType;
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.PropertyAccessor;
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.SerializationFeature;
-import com.vaadin.flow.component.ClickEvent;
 import com.vaadin.flow.component.Component;
 import com.vaadin.flow.component.ComponentEvent;
-import com.vaadin.flow.component.ComponentEventListener;
 import com.vaadin.flow.component.button.Button;
 import com.vaadin.flow.component.charts.Chart;
 import com.vaadin.flow.component.charts.demo.AbstractChartExample;
@@ -34,7 +25,6 @@ import com.vaadin.flow.component.html.Span;
 import com.vaadin.flow.component.orderedlayout.HorizontalLayout;
 import com.vaadin.flow.component.orderedlayout.VerticalLayout;
 import com.vaadin.flow.component.radiobutton.RadioButtonGroup;
-import com.vaadin.flow.server.Command;
 
 @SkipFromDemo
 public class ServerSideEvents extends AbstractChartExample {
@@ -48,6 +38,7 @@ public class ServerSideEvents extends AbstractChartExample {
     private DataSeriesItem firstDataPoint;
     private Checkbox visibilityToggling;
     private boolean setExtremes = true;
+    private boolean hideSeries = true;
 
     @Override
     public void initDemo() {
@@ -141,21 +132,12 @@ public class ServerSideEvents extends AbstractChartExample {
                 .setVisibilityTogglingDisabled(visibilityToggling.getValue()));
         visibilityToggling.setValue(false);
 
-        final Button firstSeriesVisible = new Button("Hide first series");
+        final Button firstSeriesVisible = new Button("Hide first series", e -> {
+            Series firstSeries = chart.getConfiguration().getSeries().get(0);
+            ((AbstractSeries) firstSeries).setVisible(!hideSeries);
+            hideSeries = !hideSeries;
+        });
         firstSeriesVisible.setId("hideFirstSeries");
-        firstSeriesVisible.addClickListener(
-                new ComponentEventListener<ClickEvent<Button>>() {
-                    private boolean hideSeries = true;
-
-                    @Override
-                    public void onComponentEvent(
-                            ClickEvent<Button> buttonClickEvent) {
-                        Series firstSeries = chart.getConfiguration()
-                                .getSeries().get(0);
-                        ((AbstractSeries) firstSeries).setVisible(!hideSeries);
-                        hideSeries = !hideSeries;
-                    }
-                });
 
         final RadioButtonGroup<Dimension> zoomLevels = new RadioButtonGroup<>();
         zoomLevels.setItems(Dimension.XY, Dimension.X, Dimension.Y);
@@ -215,29 +197,6 @@ public class ServerSideEvents extends AbstractChartExample {
         eventDetails.setText(details);
         Span history = new Span(name + ": " + details + "\n");
         history.setId("event" + eventNumber++);
-        historyLayout.getElement().insertChild(0, history.getElement());
-    }
-
-    private String createEventString(ComponentEvent<Chart> event) {
-        ObjectMapper mapper = new ObjectMapper()
-                .setSerializationInclusion(JsonInclude.Include.NON_NULL)
-                .disable(SerializationFeature.FAIL_ON_EMPTY_BEANS)
-                .enable(SerializationFeature.INDENT_OUTPUT)
-                .setVisibility(PropertyAccessor.ALL,
-                        JsonAutoDetect.Visibility.NONE)
-                .setVisibility(PropertyAccessor.FIELD,
-                        JsonAutoDetect.Visibility.ANY)
-                .addMixIn(Command.class, JacksonMixinForIgnoreCommand.class);
-
-        try {
-            return mapper.writeValueAsString(event);
-        } catch (JsonProcessingException e) {
-            e.printStackTrace();
-            return "";
-        }
-    }
-
-    @JsonIgnoreType
-    static class JacksonMixinForIgnoreCommand {
+        historyLayout.addComponentAtIndex(0, history);
     }
 }
